+++ src/Core/WordPress.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace MVRK\Icenberg\Core;
-
-class WordPress
-{
-}